Some style changes

+++ app/controllers/document.js
@@ -3,22 +3,23 @@ import Ember from 'ember';
 export default Ember.ObjectController.extend({
   previewVisible: true,
 
-  getCurrentDocument: function() {
-    return this.store.find('document', this.get('model.id'));
-  },
-
   actions: {
     deleteDocument: function(model) {
-      var self = this;
-      model.destroyRecord().then(function() {
-        self.transitionToRoute('documents');
-      });
+      var self = this,
+          name = this.get('name'),
+          message = 'Do you really want to delete the document "' + name + '"?';
+
+      if (confirm(message)) {
+        model.destroyRecord().then(function() {
+          self.transitionToRoute('documents');
+        });
+      }
     },
 
     saveDocument: function() {
       var self = this;
 
-      this.getCurrentDocument()
+      this.store.find('document', this.get('model.id'))
         .then(function(record) {
           return record.setProperties({name: self.get('name'), text: self.get('text')});
         }).then(function(record) {
@@ -30,6 +31,7 @@ export default Ember.ObjectController.extend({
 
     cancelDocument: function() {
       var cachedModel = this.get('cachedModel');
+
       this.setProperties({name: cachedModel.name, text: cachedModel.text});
       this.set('previewVisible', true);
     },
